requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
geoalchemy2==0.14.2
GeoAlchemy2==0.14.2
geopandas==0.14.1
pandas==2.1.3
shapely==2.0.2
fiona==1.9.5
python-dotenv==1.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
